refactor(slashChange commands): Remove duplicate users from contract order.
Removed duplicate users from the contract order to ensure each user is only listed once.
